src/main/java/com/example/Aerolinea/controller/UserController.java
package com.example.Aerolinea.controller;

import com.example.Aerolinea.model.User;
import com.example.Aerolinea.service.UserService;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping("/api/user")
@CrossOrigin
public class UserController {
    private final UserService userService;

    public UserController(UserService userService) {
        this.userService = userService;
    }


    @PostMapping
    public User createUser(@RequestBody User user) {
        return userService.createUser(user);
    }

    @GetMapping
    public List<User> getAllUsers() {
        return userService.getAllUsers();
    }

    @GetMapping("/{id}")
    public Optional<User> getUserById(@PathVariable long id) {
        return userService.getUserById(id);
    }

    @PutMapping("/{id}")
    public User updateUser(@RequestBody User user, @PathVariable long id) {
        return userService.updateUser(user, id);
    }

    @DeleteMapping("/{id}")
    public String deleteUserById(@PathVariable long id) {
        boolean ok = userService.deleteUser(id);
        if (ok) {
            return "User with id " + id + " was deleted.";
        } else {
            return "User with id " + id + " not found.";
        }
    }
}